[#79011] Chartfield1 value is not populating in XML upload
There were two version of the journal XML template, one in the main app
and one in the jxml engine.  The front end was using the app/views
template, while the actual nightly job is using the engine version.

We should use the engine version. I’ve created a more generic template
in the main app that should be pulled up into open.