Fix JXML Renderer specs and refactor renderer
We use time cop to set the time for all tests in spec_helper a year out
from the beginning of the fiscal year. When the fiscal year rolled over
on 9/1, that date became a weekend so the JxmlRenderer wouldn’t run in
the tests that assumed it would.

I also did a good amount of refactoring of the renderer itself.

Closes #87