[#103993] Accept Payments via Cardconnect
Sets up a new `nu_cardconnect` engine which allows NU to accept credit
cards via their API. We went through a lot of iterations, including
initially trying their hosted payment page. There is quite a bit of
history within the PR that could be relevant at a later time.

We're using the AJAX tokenizer, which makes an AJAX call to their server
and returns a token. That is what is passed to our server. We should never
see sensitive credit card information, and thus stay outside PCI scope.

Statements are looked up by a UUID. Anyone with the link may access the
payment page and pay, even if they are not logged in.

The merchant ID is configured on a per-facility basis from the Admin tab.

Add payment link to PDF -- Because this is NU specific already, I
figured it wasn’t worth figuring out how to add view hooks into the statement.

CC number and token are filtered from the logs in production.

The cc_number parameter could get sent to the server is javascript
breaks down. This will prevent it from appearing in the logs if it does.

Closes #105