Better handling of validator errors during expiration setting
Errors in NU such as BlacklistError were getting raised all the way up
to the view as opposed to caught and displayed in a pretty error
message. This changes the behavior so that `set_expires_at` will never
raise its own errors (I’m not clear why it did before).