Add Rollbar capistrano variables
Deployment fails without this. If we don’t set `ROLLBAR_ACCESS_TOKEN`,
then it just means that rollbar doesn’t get deployment information. We
can’t use `Settings` because it isn’t loaded at this point. Usually
this env variable is set in Circle, but we’re not using it for
deployments currently.