[#125686] Add timeout to cardconnect tokenizer (#147)
We ran into a problem where the corporate firewall was preventing
connection to port 8443, where the endpoint is served from, and the
page was hanging. This adds a three second timeout and presents an
error to the user.

I also tested this when the page is served, but then internet
connection is lost. That provides a `errorThrown == “error”`. We’ll
throw a JS error in that case.