Capture encoding errors on order import
If there is an invalid character in a CSV, we would get lots of error
notification. This captures those errors and provides a more useful
message to the user. In Oracle, it would raise an error, but MySQL
would eat the error and just not find the account or product. By adding
the `encode` at the app layer, it makes sure the error gets raised.